Cap max version for sqlalchemy in Wallaby
CentOS Stream 9 has recently added python-sqlalchemy-1.4.37. OpenStack
Wallaby has version 1.3.23 in upper-constraints and it's not compatible
with 1.4 as this version introduced some non-backwards compatible
changes.

This change forces to install sqlalchemy < 1.4.0 so that it pulls
python-sqlalchemy13 to resolve the requirement.

Also, it backports usptream patch [1] required to run unit tests with
python3 and current version of testtools.

[1] https://review.opendev.org/c/openstack/networking-sfc/+/791359

Change-Id: I90a9380379a0e92e21380872d58d974f08894c15